+++ block/src/lib.rs
@@ -14,12 +14,6 @@ pub mod factory;
 #[path = "io/mod.rs"]
 mod io_impl;
 pub use io_impl::{async_io, fcntl, request};
-pub mod fixed_vhd;
-#[cfg(feature = "io_uring")]
-/// Enabled with the `"io_uring"` feature
-pub mod fixed_vhd_async;
-pub mod fixed_vhd_disk;
-pub mod fixed_vhd_sync;
 pub mod formats;
 pub mod qcow;
 #[cfg(feature = "io_uring")]
@@ -29,7 +23,6 @@ pub mod qcow_disk;
 pub(crate) mod qcow_sync;
 mod sparse;
 pub use formats::raw as raw_disk;
-pub mod vhd;
 pub mod vhdx;
 pub mod vhdx_sync;
 
@@ -521,7 +514,7 @@ pub fn detect_image_type(f: &mut File) -> BlockResult<ImageType> {
     // Check 4 first bytes to get the header value and determine the image type
     let image_type = if u32::from_be_bytes(block[0..4].try_into().unwrap()) == QCOW_MAGIC {
         ImageType::Qcow2
-    } else if vhd::is_fixed_vhd(f)
+    } else if formats::vhd::is_fixed_vhd(f)
         .map_err(|e| BlockError::new(BlockErrorKind::Io, e).with_op(ErrorOp::DetectImageType))?
     {
         ImageType::FixedVhd
